+++ include/yolocpp/engine/trainer.hpp
@@ -20,11 +20,15 @@
 #include <vector>
 
 #include "yolocpp/datasets/yolo_dataset.hpp"
+#include "yolocpp/losses/yolo1_loss.hpp"
+#include "yolocpp/losses/yolo2_loss.hpp"
 #include "yolocpp/losses/yolo10_loss.hpp"
 #include "yolocpp/losses/yolo26_loss.hpp"
 #include "yolocpp/losses/yolo6_loss.hpp"
 #include "yolocpp/losses/yolo7_loss.hpp"
 #include "yolocpp/losses/yolo8_loss.hpp"
+#include "yolocpp/models/yolo1.hpp"
+#include "yolocpp/models/yolo2.hpp"
 #include "yolocpp/models/yolo11.hpp"
 #include "yolocpp/models/yolo12.hpp"
 #include "yolocpp/models/yolo13.hpp"
@@ -120,5 +124,7 @@ using TrainerV11 = TrainerT<models::Yolo11Detect>;
 using TrainerV12 = TrainerT<models::Yolo12Detect>;
 using TrainerV13 = TrainerT<models::Yolo13Detect>;
 using TrainerV26 = TrainerT<models::Yolo26Detect>;
+using TrainerV1  = TrainerT<models::Yolo1>;
+using TrainerV2  = TrainerT<models::Yolo2>;
 
 }  // namespace yolocpp::engine