+++ include/yolocpp/engine/validator.hpp
@@ -8,6 +8,8 @@
 #include "yolocpp/datasets/yolo_dataset.hpp"
 #include "yolocpp/inference/nms.hpp"
 #include "yolocpp/metrics/map.hpp"
+#include "yolocpp/models/yolo1.hpp"
+#include "yolocpp/models/yolo2.hpp"
 #include "yolocpp/models/yolo11.hpp"
 #include "yolocpp/models/yolo12.hpp"
 #include "yolocpp/models/yolo13.hpp"